+++ singer-connectors/tap-postgres/tests/unit/test_full_table.py
@@ -14,6 +14,7 @@ class TestFullTable(TestCase):
         super(TestFullTable, cls).setUpClass()
         cls.patcher = patch('psycopg2.connect')
         mocked_connect = cls.patcher.start()
+        mocked_connect.return_value.server_version = 110002
         mocked_connect.return_value.__enter__.return_value = MockedConnect()
 
     @classmethod